refactor(core): move BlockStore and BlockRegistry into src/store/
