+++ packages/core/src/store/BlockStore.ts
@@ -1,7 +1,7 @@
-import {signal} from '../signals'
-import type {DragAction, DragActions} from '../types'
-import {getDragDropPosition, getDragTargetIndex, parseDragSourceIndex} from '../utils/dragUtils'
-import {isClickOutside, isEscapeKey} from '../utils/menuUtils'
+import {signal} from '../shared/signals'
+import type {DragAction, DragActions} from '../shared/types'
+import {getDragDropPosition, getDragTargetIndex, parseDragSourceIndex} from '../shared/utils/dragUtils'
+import {isClickOutside, isEscapeKey} from '../shared/utils/menuUtils'
 
 export type DropPosition = 'before' | 'after' | null
