+++ packages/core/src/store/index.ts
@@ -1,1 +1,4 @@
 export {Store} from './Store'
+export {BlockStore} from './BlockStore'
+export type {DropPosition} from './BlockStore'
+export {BlockRegistry} from './BlockRegistry'
